requirements.txt
langchain
llama-index
sentence-transformers
chromadb
openai
tiktoken
python-dotenv
-e .
